pyproject.toml
[tool.poetry]
name = "netops.io"
version = "0.4.0"
description = "netops.io web application and rest framework"
authors = ["James Williams <james.williams@packetgeek.net>"]
license = "Apache"

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.7"
djangorestframework = "^3.13.1"
celery = {extras = ["redis"], version = "^5.2.1"}
invoke = "^1.6.0"
django-celery-results = "^2.2.0"
mysqlclient = "^2.1.0"
uritemplate = "^4.1.1"
drf-yasg = {extras = ["validation"], version = "^1.20.0"}
django-crispy-forms = "^1.14.0"
crispy-bootstrap5 = "^0.6"

[tool.poetry.dev-dependencies]
black = "^21.8b0"
pytest = "^6.2.5"
pylint = "^2.10.2"
bandit = "^1.7.0"
flake8 = "^3.9.2"
coverage = "^6.4.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
